refactor(p2): extract AskUserToolStep into sibling (stage 2 of 4)
Moves the AskUser tool-call step Composable plus its helpers (AskUserToolStep,
AskOptionChip, AskUserQuestion data class, ToolDenyReasonDialog) from
ChatMessageTools.kt L1239-L1644 (parent ae597a0d) into a new sibling
ChatMessageAskUserStep.kt. Bodies are byte-equal with the parent commit.

Widening:
  - sibling: AskUserToolStep private → internal (called from
    ChatMessageToolStep L566 in coordinator).
  - sibling: ToolDenyReasonDialog private → internal (called from
    ChatMessageToolStep L883 in coordinator).
  - AskOptionChip and AskUserQuestion data class stay private (intra-block
    only).

The block is fully self-contained against coordinator helpers: zero refs to
ToolNames / MemoryActions / getStringContent / AgentToolStatus / AgentToolKind
/ AgentToolCallCapsule etc. Only external receiver type used is
ChainOfThoughtScope (imported).

Two `by` delegations in the block (var expanded, var reason) — sibling imports
both getValue + setValue explicitly per playbook §5.3 (these are not literal
grep matches inside the block body).

Coordinator strip: 2254 → 1857 (-397 body) → 1845 (-12 import cleanup).
Imports trimmed: 12 this-refactor orphans (ExperimentalLayoutApi, FlowRow,
PaddingValues, AlertDialog, Button, OutlinedTextField, OutlinedTextFieldDefaults,
TextButton, mutableStateMapOf, buildJsonObject, buildJsonArray, DotLoading).
Pre-existing dead imports preserved (CLAUDE.md §3): Box, BoxWithConstraints,
CircleShape, FilledTonalButton, FilterChip, LocalContentColor, Color, getValue,
setValue (last two used implicitly via `by` — 14 sites in stripped coord).

Baseline 467 tests / 3 failed preserved.
